:fire: chore: remove all the dead code
