+++ core/package.json
@@ -51,11 +51,11 @@
     "wagmi": "^1.3.0",
     "zod": "^3.21.4"
   },
+  "publishConfig": {
+    "access": "public"
+  },
   "optionalPeerDependencies": {
     "wagmi": ">1.0.0",
     "viem": ">1.0.0"
-  },
-  "publishConfig": {
-    "access": "public"
   }
 }
