--- core/src/hardFork.ts
@@ -1,6 +1,0 @@
-import { Chain, Common, Hardfork } from '@ethereumjs/common'
-
-export const common = new Common({
-	chain: Chain.Rinkeby,
-	hardfork: Hardfork.Istanbul,
-})
